tests: improve test setup for spying tracer
